Fix Next.js 15 Suspense boundary error for useSearchParams in admin login page
