+++ components/AdminLoginForm.tsx
@@ -1,0 +1,130 @@
+'use client'
+
+import { useState } from 'react'
+import { useRouter, useSearchParams } from 'next/navigation'
+import { Lock, Eye, EyeOff } from 'lucide-react'
+
+export default function AdminLoginForm() {
+  const [accessCode, setAccessCode] = useState('')
+  const [showPassword, setShowPassword] = useState(false)
+  const [error, setError] = useState('')
+  const [isLoading, setIsLoading] = useState(false)
+  
+  const router = useRouter()
+  const searchParams = useSearchParams()
+  const redirectPath = searchParams.get('redirect') || '/bookings'
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault()
+    setError('')
+    setIsLoading(true)
+
+    try {
+      // Verify the access code by making a request with it
+      const response = await fetch('/api/verify-access', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ accessCode }),
+      })
+
+      if (response.ok) {
+        // Redirect with access code as query parameter to set cookie
+        const redirectUrl = new URL(redirectPath, window.location.origin)
+        redirectUrl.searchParams.set('access_code', accessCode)
+        window.location.href = redirectUrl.toString()
+      } else {
+        setError('Invalid access code. Please try again.')
+      }
+    } catch (error) {
+      console.error('Access code verification error:', error)
+      setError('An error occurred. Please try again.')
+    } finally {
+      setIsLoading(false)
+    }
+  }
+
+  return (
+    <div className="min-h-screen bg-gray-50 flex items-center justify-center px-4">
+      <div className="max-w-md w-full space-y-8">
+        <div className="text-center">
+          <div className="mx-auto h-12 w-12 bg-blue-100 rounded-full flex items-center justify-center">
+            <Lock className="h-6 w-6 text-blue-600" />
+          </div>
+          <h2 className="mt-6 text-3xl font-bold text-gray-900">
+            Admin Access Required
+          </h2>
+          <p className="mt-2 text-sm text-gray-600">
+            Enter your access code to manage bookings and settings
+          </p>
+        </div>
+
+        <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
+          <div>
+            <label htmlFor="access-code" className="sr-only">
+              Access Code
+            </label>
+            <div className="relative">
+              <input
+                id="access-code"
+                name="access-code"
+                type={showPassword ? 'text' : 'password'}
+                required
+                className="appearance-none rounded-lg relative block w-full px-3 py-3 pl-4 pr-12 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 focus:z-10 text-lg"
+                placeholder="Enter access code"
+                value={accessCode}
+                onChange={(e) => setAccessCode(e.target.value)}
+                disabled={isLoading}
+              />
+              <button
+                type="button"
+                className="absolute inset-y-0 right-0 pr-3 flex items-center"
+                onClick={() => setShowPassword(!showPassword)}
+                disabled={isLoading}
+              >
+                {showPassword ? (
+                  <EyeOff className="h-5 w-5 text-gray-400 hover:text-gray-600" />
+                ) : (
+                  <Eye className="h-5 w-5 text-gray-400 hover:text-gray-600" />
+                )}
+              </button>
+            </div>
+          </div>
+
+          {error && (
+            <div className="bg-red-50 border border-red-200 rounded-lg p-3">
+              <p className="text-sm text-red-600">{error}</p>
+            </div>
+          )}
+
+          <div>
+            <button
+              type="submit"
+              disabled={isLoading || !accessCode}
+              className="group relative w-full flex justify-center py-3 px-4 border border-transparent text-sm font-medium rounded-lg text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:bg-gray-400 disabled:cursor-not-allowed transition-colors duration-200"
+            >
+              {isLoading ? (
+                <div className="flex items-center">
+                  <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
+                  Verifying...
+                </div>
+              ) : (
+                'Access Admin Panel'
+              )}
+            </button>
+          </div>
+
+          <div className="text-center">
+            <a
+              href="/"
+              className="text-sm text-blue-600 hover:text-blue-500 transition-colors duration-200"
+            >
+              ← Back to Home
+            </a>
+          </div>
+        </form>
+      </div>
+    </div>
+  )
+}
